chore: placement — rollback outcome and pack provenance codec get own files
The compensation collector moves out of the provenance data model into
api/registry/rollback.go; the pack provenance encode/decode pair and the
file digest live together in cmd/wippy/cmd/pack_provenance.go.
